RAG_PROJECT/requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langchain-ollama
faiss-cpu
tiktoken
python-dotenv
